pyproject.toml
[project]
name = "weclaw"
version = "0.1.0"
description = "Mini OpenClaw built on the Python LangChain framework"
requires-python = ">=3.10"
dependencies = [
    "deepagents",
    "langchain",
    "langchain-community",
    "langchain-ollama",
    "langchain-anthropic",
    "langchain-google-genai",
    "langchain-core",
    "langchain-openai",
    "openai",
    "dashscope",
    "python-dotenv",
    "PyYAML",
    "mcp",
    "python-socks",
    "python-telegram-bot",
    "websockets",
    "aiosqlite",
    "langgraph-checkpoint-sqlite",
    "apscheduler",
    "httpx",
    "lark-oapi",
    "certifi",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
claw = "weclaw.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weclaw"]

[tool.ruff]
target-version = "py310"
line-length = 120
